🧹 moved generic components from gurb to common
