+++ src/containers/GurbForm.jsx
@@ -12,9 +12,9 @@ import NewMember from './Gurb/NewMember'
 import Contract from './Gurb/Contract'
 import Gurb from './Gurb/Gurb'
 
-import PrevButton from './Gurb/components/PrevButton'
-import NextButton from './Gurb/components/NextButton'
-import SubmitButton from './Gurb/components/SubmitButton'
+import PrevButton from '../components/NewButtons/PrevButton'
+import NextButton from '../components/NewButtons/NextButton'
+import SubmitButton from '../components/NewButtons/SubmitButton'
 import supplyPointValidations from './Gurb/supplyPointValidations'
 import {
   addressValidations,
